requirements/base.txt
Django==1.9.8
django-environ==0.4.3
djangorestframework==3.3.2
requests==2.18.4
phonenumbers==8.5.0
gunicorn==19.7.1
mysqlclient==1.3.10
pytz==2016.10
python-dateutil==2.6.1